Create example of configuring a marker's anchor
The ability to configure a marker's anchor point using Google Maps is
less than intuitive. Although the documentation makes it sound like you
can configure it via a MarkerOptions object, that doesn't seem to work
with this library.  As such, you need to configure the anchor point on
the image itself.  This example demonstrates how to do that.