Fix Performance issues on drag and zoom
Scope variables like center and zoom are updated on every drag / zoom event. Fixed that by changing _.defer to _.debounce with a configuration debounce delay.

See https://github.com/angular-ui/angular-google-maps/issues/734

Please test this change, it solves the issues in my case (made the changes in JS), but I wrote the coffeescript without being able to test it, as I could not get the project to build locally.