Fix for:  icons to be rendered along the polyline.
Describes how icons are to be rendered on a line. 

If your polyline is geodesic, then the distances specified for both offset and repeat are calculated in meters by default. Setting either offset or repeat to a pixel value will cause the distances to be calculated in pixels on the screen.